requirements.txt
fastapi
uvicorn
pillow
opencv-python-headless
numpy
python-multipart